Refactoring of '_parent_of' function to accept a inheritance depth.
The idea is to test recursively against the catalog pg_inherits
to check for an inheritance chain between two tables. For a direct relationship,
the depth is 1 (default value), for deeper inheritance chains it must be
specified.
Since this uses a recursive CTE, it could have some problems in porting
to old PostgreSQL versions.