Unindent subtest diag header
In order to make it clearer where a subest starts. This contrasts with
how Test::More does it, but is easier to read for errors and aligns
with how some other TAP frameworks like node tap do it. The format
change will have no impact on how TAP-parsing tools like `pg_prove`
interpret test output, but could break diff-based tooling such as
`pg_regress`. That change is demonstrated in the expected files
modified in this commit.