Switch plan(int) to plan(bigint)
A useful test pattern is to define a matrix of tests in a table or view
by building a cartesian join of multiple dimensions. If you do that,
you'll want to do something along the lines of

SELECT plan(
  0
  + <some number of static tests>
  + (SELECT count(*) FROM test_view)
);

Because count(*) returns bigint doing this with plan(int) forces you to
cast the output of count(*) to int. Switching plan() to accept a bigint
removes the need for that cast.